✨ Made map component with filter a separate component :bug: Fixed bug on regions
